sockutil: fix derp

+++ src/sockutil.h
@@ -21,7 +21,7 @@ typedef union {
 bool socket_set_nonblock(int fd);
 void socket_set_reuseport(int fd, bool reuseport);
 void socket_set_tcp(int fd, bool nodelay, bool keepalive);
-void socket_set_buffer(int fd, size_t send, size_t recv);
+void socket_set_buffer(int fd, int send, int recv);
 void socket_bind_netdev(int fd, const char *netdev);
 int socket_get_error(int fd);
